Added some Table Data Grabbers
No tests written just yet, exact structure for the <td> data grabbers will be decided as they are implemented. Tests will be written so that they contain the html <tr> entry and desired results, then the grabbers will be tested on if they return what is correct.

Issue can be removed as data grabbers implemented are returning lists to address the 'multiples issue' of the multiple sessions on a single <tr>