Added all table row table data fetchers.
From here, organization of the fetchers will be done. They are distinct pieces of functionality, but, they all are ran at the same time for every table row on the results page. Perhaps a class will be made to read the information of a table row and form it into an entry.